+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added `VariableScope~variables` to access all variables as a plain object
 * Ensure that `Xyz.isXyz()` functions always return a boolean
 * Fixed a bug which caused `Request` to not initialize headers
+* Added generic `.toObject` and `.toString` method to PropertyList to work for Types that has a .valueOf and `.unparse` defined.
+* Added HeaderList and **discontinued** `Header.headerValue` in favour of `HeaderList.get`
 
 #### 1.1.0 (April 03, 2017)
 * Enhanced the `PropertyList` to allow keys with multiple values