+++ CHANGELOG.md
@@ -5,7 +5,13 @@
 * Added `VariableScope~variables` to access all variables as a plain object
 * Ensure that `Xyz.isXyz()` functions always return a boolean
 * Fixed a bug which caused `Request` to not initialize headers
-* Added generic `.toObject` and `.toString` method to PropertyList to work for Types that has a .valueOf and `.unparse` defined.
+* `RequestAuthBase` inherits from `PropertyBase`
+* All auth definitions inherit from `RequestAuthBase` (through a dynamic intermediate class)
+* Added `Item.getAuth()` which returns an instance of the currently selected auth-type (searches parent folders as well)
+* Added `PropertyBase.findInParents` which finds a given property in the closest parent.
+* Added `PropertyBase.findParentContaining` which finds the closest parent containing the given property.
+* Added `RequestAuth.current()` which returns an instance of the currently selected auth.
+* Added `.toObject` and `.toString` method to PropertyList to work for Types that has a .valueOf and `.unparse` defined.
 * Added HeaderList and **discontinued** `Header.headerValue` in favour of `HeaderList.get`
 
 #### 1.1.0 (April 03, 2017)