+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 * Fixed issue with Cookie constructor misbehaving when constructor definition was missing
 * Fixed issue with `Cookie~update` function causing previosuly defined properties from being overwritten
 * Updated Cookie to behave properly in lists (`name` is now the key and it is multi-value case insensitive)
+* `RequestAuthBase` inherits from `PropertyBase`
+* All auth definitions inherit from `RequestAuthBase` (through a dynamic intermediate class)
+* Added `Item.getAuth()` which returns an instance of the currently selected auth-type (searches parent folders as well)
+* Added `PropertyBase.findInParents` which finds a given property in the closest parent.
+* Added `PropertyBase.findParentContaining` which finds the closest parent containing the given property.
+* Added `RequestAuth.current()` which returns an instance of the currently selected auth.
 
 #### 1.1.0 (April 03, 2017)
 * Enhanced the `PropertyList` to allow keys with multiple values